refactor(gcf-utils): extract Octokit request parsing logic to separate module (#769)
The Octokit logging plugin in gcf-utils has some great request parsing logic that should be reusable. Hence pulling that logic out to a separate module.
